Read model_dtype and autocast_dtype from validate_model_defaults.yaml

+++ examples/compress/main.py
@@ -32,8 +32,6 @@ import argparse
 from datetime import timedelta
 from pathlib import Path
 
-import torch
-
 import modelopt.torch._compress.mip.mip_and_realize_models as mip_and_realize_models
 import modelopt.torch.nas as mtn
 import modelopt.torch.utils.distributed as dist
@@ -148,9 +146,7 @@ def run_mip_only(hydra_config_path: str):
     # mip_and_realize_models (distributed processing)
     # TODO: How to make it part of mnt.search() api, similarly to run_full_compress() API
     mprint("Compress Progress 7/8: running MIP and realizing models (multi-gpu)")
-    mip_and_realize_models.launch_mip_and_realize_model(
-        hydra_cfg, model_dtype=torch.bfloat16, autocast_dtype=torch.bfloat16
-    )
+    mip_and_realize_models.launch_mip_and_realize_model(hydra_cfg)
 
     dist.cleanup()
     mprint("Compress Progress 8/8: compression pipeline completed (multi-gpu)")
